Refactor in topsidebar,commonactions and fixes
