+++ test/pages/contacts/contactsBaseForm.page.js
@@ -1,5 +1,5 @@
 let commonActions = require("../../utils/CommonActions");
-let Form = require('../form.page');
+let Form = require('../common/form.page');
 
 class ContactsBaseForm extends Form {
     /**
